Developer

+++ src/db/trades.js
@@ -1,4 +1,5 @@
-import Trade from "../models/Trade.js";
+import Trade, { STATUS_LEVEL } from "../models/Trade.js"
+import Plant from "../models/Plant.js";
 
 export async function getAllTrades() {
   try {
@@ -36,24 +37,37 @@ export async function createTrade(tradeData) {
 
 export async function updateTrade(id, tradeData) {
   try {
-    const updatedTrade = await Trade.findById(id);
+    const updatedTrade = await Trade.findById(id)
 
+    if (!updatedTrade) return null
+       if (tradeData.status === STATUS_LEVEL.cancelled) {
+        if (updatedTrade.status !== STATUS_LEVEL.pending && updatedTrade.status !== STATUS_LEVEL.approved) {
+          throw new Error("Trade can only be cancelled if status is pending")
+        }
+        await Plant.findByIdAndUpdate(updatedTrade.plantId, { available: true })
+        await Trade.findByIdAndDelete(id)
+        return { cancelled:true }
+       }
+      
     updatedTrade.status = tradeData.status ?? updatedTrade.status;
-
     await updatedTrade.save();
-
-    console.log("Dunction: updateTrade:", updateTrade);
-    
+    // console.log("Function: updateTrade:", updateTrade);
     return await Trade.populate(updatedTrade, "plantId requesterId ownerId");
   } catch (err) {
-    console.error("Unable to update 'Trade'", err);
+    console.error("Unable to update 'Trade'", err)
+    throw err;
   }
 }
 
 export async function deleteTrade(id) {
-  try {
+ try {
+    const trade = await Trade.findById(id);
+    if (!trade) return false; 
+    
+  
     return !!(await Trade.findByIdAndDelete(id));
   } catch (err) {
-    console.error("Unable to delete 'Trade'", err);
+    console.error("Unable to delete 'Trade'", err)
+    return false; 
   }
 }
